+++ src/main/java/com/example/bookrating/controller/ReviewController.java
@@ -14,22 +14,22 @@ public class ReviewController {
     private ReviewService reviewService;
 
     @GetMapping("/{bookId}/reviews")
-    public ReviewListResponseDto getReviews(@PathVariable("bookId") Integer bookId) {
+    public ReviewListResponseDto getReviews(@PathVariable("bookId") Long bookId) {
         return reviewService.getReviews(bookId);
     }
 
     @PostMapping("/{bookId}/reviews")
-    public ReviewResponseDto addReview(@PathVariable("bookId") Integer bookId, @RequestBody ReviewDto dto) {
+    public ReviewResponseDto addReview(@PathVariable("bookId") Long bookId, @RequestBody ReviewDto dto) {
         return reviewService.addReview(bookId, dto);
     }
 
     @PatchMapping("/{bookId}/reviews/{reviewId}")
-    public ReviewResponseDto updateReview(@PathVariable("bookId") Integer bookId, @PathVariable("reviewId") Long reviewId, @RequestBody ReviewDto dto) {
+    public ReviewResponseDto updateReview(@PathVariable("bookId") Long bookId, @PathVariable("reviewId") Long reviewId, @RequestBody ReviewDto dto) {
         return reviewService.updateReview(bookId, reviewId, dto);
     }
 
     @DeleteMapping("/{bookId}/reviews/{reviewId}")
-    public ReviewResponseDto deleteReview(@PathVariable("bookId") Integer bookId, @PathVariable("reviewId") Long reviewId) {
+    public ReviewResponseDto deleteReview(@PathVariable("bookId") Long bookId, @PathVariable("reviewId") Long reviewId) {
         return reviewService.deleteReview(bookId, reviewId);
     }
 }
