+++ src/main/java/com/example/bookrating/service/ReviewService.java
@@ -21,7 +21,7 @@ public class ReviewService {
     @Autowired
     private BookService bookService;
 
-    public ReviewListResponseDto getReviews(int bookId) {
+    public ReviewListResponseDto getReviews(Long bookId) {
         bookService.findBookOrThrow(bookId);
         List<ReviewDto> reviewDtos = reviewRepository.findReviewByBookId(bookId).stream()
                 .map(ReviewService::getReviewDto)
@@ -31,7 +31,7 @@ public class ReviewService {
         return new ReviewListResponseDto(reviewDtos, averageRating);
     }
 
-    private double getAverageRating(int bookId) {
+    private double getAverageRating(Long bookId) {
         Double averageRating = reviewRepository.findAverageRatingByBookId(bookId).orElse(0.0);
         return BigDecimal.valueOf(averageRating)
                 .setScale(2, RoundingMode.HALF_UP)
@@ -42,13 +42,13 @@ public class ReviewService {
         return new ReviewDto(review.getId(), review.getRating(), review.getReviewText(), review.getUpdatedAt());
     }
 
-    public ReviewResponseDto addReview(int bookId, ReviewDto dto) {
+    public ReviewResponseDto addReview(Long bookId, ReviewDto dto) {
         Book book = bookService.findBookOrThrow(bookId);
         Review saved = reviewRepository.save(new Review(book, dto.getRating(), dto.getReviewText()));
         return new ReviewResponseDto(getReviewDto(saved), getAverageRating(bookId));
     }
 
-    public ReviewResponseDto updateReview(int bookId, Long reviewId, ReviewDto dto) {
+    public ReviewResponseDto updateReview(Long bookId, Long reviewId, ReviewDto dto) {
         Book book = bookService.findBookOrThrow(bookId);
 
         Review review = findReviewOrThrow(reviewId);
@@ -63,7 +63,7 @@ public class ReviewService {
         return reviewRepository.findById(reviewId).orElseThrow(() -> new IllegalStateException("존재하지 않는 리뷰입니다"));
     }
 
-    public ReviewResponseDto deleteReview(int bookId, Long reviewId) {
+    public ReviewResponseDto deleteReview(Long bookId, Long reviewId) {
         bookService.findBookOrThrow(bookId);
         findReviewOrThrow(reviewId);
         reviewRepository.deleteById(reviewId);
